requirements.txt
click
pillow
pysvg-py3
